// AngularJS/src/ngLanguageService/typings/types.d.ts
export interface Span {
    /**
     * The first code-point of the span as an offset relative to the beginning of the source assuming
     * a UTF-16 encoding.
     */
    start: number;
    /**
     * The first code-point after the span as an offset relative to the beginning of the source
     * assuming a UTF-16 encoding.
     */
    end: number;
}
// ...
export interface Completion {
    /**
     * The kind of comletion.
     */
    kind: CompletionKind;
    /**
     * The name of the completion to be displayed
     */
    name: string;
    /**
     * The key to use to sort the completions for display.
     */
    sort: string;
}
/**
 * A sequence of completions.
 *
 * @experimental
 */
export declare type Completions = Completion[];
/**
 * A file and span.
 */
export interface Location {
    fileName: string;
    span: Span;
}
/**
 * A defnition location(s).
 */
export declare type Definition = Location[];
/**
 * The kind of diagnostic message.
 *
 * @experimental
 */
export declare enum DiagnosticKind {
    Error = 0,
    Warning = 1,
}
/**
 * An template diagnostic message to display.
 *
 * @experimental
 */
export interface Diagnostic {
    /**
     * The kind of diagnostic message
     */
    kind: DiagnosticKind;
    /**
     * The source span that should be highlighted.
     */
    span: Span;
    /**
     * The text of the diagnostic message to display.
     */
    message: string;
}
/**
 * A sequence of diagnostic message.
 *
 * @experimental
 */
export declare type Diagnostics = Diagnostic[];
/**
 * Information about the pipes that are available for use in a template.
 *
 * A `LanguageService` interface.
 *
 * @experimental
 */
export interface PipeInfo {
    /**
     * The name of the pipe.
     */
    name: string;
    /**
     * The static symbol for the pipe's constructor.
     */
    symbol: any;
}
/**
 * A sequence of pipe information.
 *
 * @experimental
 */
export declare type Pipes = PipeInfo[];
// ...
export interface HoverTextSection {
    /**
     * Source code or markdown text describing the symbol a the hover location.
     */
    text: string;
    /**
     * The langauge of the source if `text` is a souce code fragment.
     */
    language?: string;
}
/**
 * Hover infomration for a symbol at the hover location.
 */
export interface Hover {
    /**
     * The hover text to display for the symbol at the hover location. If the text includes
     * source code, the section will specify which langauge it should be interpreted as.
     */
    text: HoverTextSection[];
    /**
     * The span of source the hover covers.
     */
    span: Span;
}
// ...
export interface LanguageService {
    /**
     * Returns a list of all the external templates referenced by the project.
     */
    getTemplateReferences(): string[];
    /**
     * Returns a list of all error for all templates in the given file.
     */
    getDiagnostics(fileName: string): Diagnostics;
    /**
     * Return the completions at the given position.
     */
    getCompletionsAt(fileName: string, position: number): Completions;
    /**
     * Return the definition location for the symbol at position.
     */
    getDefinitionAt(fileName: string, position: number): Definition;
    /**
     * Return the hover information for the symbol at position.
     */
    getHoverAt(fileName: string, position: number): Hover;
    /**
     * Return the pipes that are available at the given position.
     */
    getPipesAt(fileName: string, position: number): Pipes;
}